restructuring protofix code to build nicely
