+++ src/pkg/protofix/protofix/example1_pb_go
@@ -1,0 +1,9 @@
+package protofix
+
+type Dummy struct {
+	NotSomethingICareAbout   string `protobuf:"bytes,2,opt,name=id" json:"id,omitempty"`
+}
+
+type Commit struct {
+	Id   string `protobuf:"bytes,2,opt,name=id" json:"id,omitempty"`
+}
